hollow out settings module
