+++ rust/gui-client/src-common/src/lib.rs
@@ -2,4 +2,5 @@ pub mod auth;
 pub mod crash_handling;
 pub mod deep_link;
 pub mod logging;
+pub mod settings;
 pub mod updates;
